Initialization of favorites adapter
-Added favorites fragment and created initial adapter for it
-Added this fragment to bottom navigation bar

+++ ecommerce/android/Getflix/app/src/main/java/com/example/getflix/model/ProductModel.kt
@@ -1,7 +1,7 @@
 package com.example.getflix.model
 
 data class ProductModel(
-    var id: Double,
+    var id: Int,
     var name: String,
     var category: String
 )
